+++ app/discover/defi/page.tsx
@@ -7,6 +7,7 @@ import {
   getLendingStats,
   getPairingStats,
 } from "@services/apiService";
+import { formatStatsData } from "@utils/defi";
 import React, { useEffect, useCallback } from "react";
 
 export default function Page() {
@@ -20,68 +21,12 @@ export default function Page() {
       const pairingStats = await getPairingStats();
       const altProtocolStats = await getAltProtocolStats();
 
-      const res: TableInfo[] = [];
-
-      if (!derivatesStats) return;
-      Object.keys(derivatesStats).map((eachKey) => {
-        const item: TableInfo = {
-          title: derivatesStats[eachKey].protocol,
-          action: "Derivatives",
-          apr: derivatesStats[eachKey].apr * 100,
-          volume: derivatesStats[eachKey].volumes,
-          daily_rewards: derivatesStats[eachKey].allocation,
-          app: eachKey,
-        };
-        res.push(item);
-      });
-
-      if (!lendingStats) return;
-      Object.keys(lendingStats).map((eachKey) => {
-        Object.keys(lendingStats[eachKey]).map((eachSubKey) => {
-          const item: TableInfo = {
-            title: eachSubKey,
-            action: "Lend",
-            apr: lendingStats[eachKey][eachSubKey].strk_grant_apr_nrs * 100,
-            volume: lendingStats[eachKey][eachSubKey].supply_usd,
-            daily_rewards: lendingStats[eachKey][eachSubKey].allocation,
-            app: eachKey,
-          };
-          res.push(item);
-        });
-      });
-
-      if (!pairingStats) return;
-      Object.keys(pairingStats).map((eachKey) => {
-        Object.keys(pairingStats[eachKey]).map((eachSubKey) => {
-          if (eachSubKey.toLocaleLowerCase() === "discretionary") return;
-          const item: TableInfo = {
-            title: eachSubKey,
-            action: "Provide Liquidity",
-            apr: pairingStats[eachKey][eachSubKey].apr * 100,
-            volume: pairingStats[eachKey][eachSubKey].tvl_usd,
-            daily_rewards: pairingStats[eachKey][eachSubKey].allocation,
-            app: eachKey,
-          };
-          res.push(item);
-        });
-      });
-
-      if (!altProtocolStats) return;
-
-      Object.keys(altProtocolStats).map((eachKey) => {
-        Object.keys(altProtocolStats[eachKey]).map((eachSubKey) => {
-          const item: TableInfo = {
-            title: eachSubKey,
-            action: "Strategies",
-            apr: altProtocolStats[eachKey][eachSubKey].apr * 100,
-            volume: altProtocolStats[eachKey][eachSubKey].tvl_usd,
-            daily_rewards: altProtocolStats[eachKey][eachSubKey].allocation,
-            app: eachKey,
-          };
-          res.push(item);
-        });
-      });
-
+      const res = formatStatsData(
+        derivatesStats,
+        lendingStats,
+        pairingStats,
+        altProtocolStats
+      );
       setData(res);
       setLoading(false);
     } catch (e) {
